pyproject.toml
[tool.poetry]
name = "electronics-sales-network-api"
version = "0.1.0"
description = ""
authors = ["Daniil Karlov <dskarlov@icloud.com>"]
readme = "README.md"
packages = [{include = "electronics_sales_network_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
psycopg2-binary = "^2.9.7"
envparse = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
